+++ CHANGELOG.md
@@ -14,12 +14,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DAML Module for Wallet.
 - DAML Module for Onboarding.
 - Changelog.
+- Merged `daml-topl-lib` and `daml-topl-broker` into one project.
 
 ### Changed
 
 - Updated DAML-SDK to version 2.7.0.
 - Updated DAML module version to 0.2.1
 - Emptied README.md to reflect the new version of the project.
+- Renamed the project to `daml-on-topl`.
+- Renamed the `daml-bifrost-module` to `daml-topl-lib`.
+- Renamed the `bifrost-daml-broker` to `daml-topl-broker`.
 
 ### Removed